Make Pexpect importable on Windows
